Fix for api.service delete function

+++ src/core/synacta/api.service.ts
@@ -96,13 +96,13 @@ export class SynactaAPIService {
     /*
      *
      */
-     private deleteBase(target: string, type: string, id: string) {
-         let endpoint = API_URL;
-         endpoint = endpoint + "base/";
-         endpoint = (type)? endpoint + type : endpoint;
-         endpoint = (id)? endpoint + "/" + id : endpoint;
-         endpoint = (target) ? endpoint + "/" + target : endpoint;
-         return this.deleteByLink(endpoint);
+     private deleteBase(type: string, id: string) {
+        if ( type == undefined || id == undefined ) {
+            return null;
+        }
+        let endpoint = API_URL;
+        endpoint += "base/" + type + "/" + id;
+        return this.deleteByLink(endpoint);
      }
 
      /*
@@ -137,10 +137,7 @@ export class SynactaAPIService {
      private deleteByLink(endpoint: string) {
          let headers = new Headers(this.baseHeaders);
          return this.http
-            .delete(endpoint, {headers: headers})
-            .retryWhen(error => error.delay(500))
-            .timeout(this.timeout, new Error("Delay Exceeded!"))
-            .map(response => response.json());
+            .delete(endpoint, {headers: headers});
      }
 
     /*
@@ -292,8 +289,8 @@ export class SynactaAPIService {
     * This function deletes a given container
     * @param container
     */
-    public deleteEntity(entity: Entity): Observable<String>{
-        return this.deleteBase(null, entity.ObjectType, entity.ID);
+    public deleteEntity(entity: Entity) {
+        return this.deleteBase(entity.ObjectType, entity.ID);
     }
 
    /* This function uses a type and an id of a document to receive
